+++ segmentation/segmentation/gpu_common.h
@@ -5,6 +5,6 @@
 #include <device_launch_parameters.h>
 
 extern "C" int mymain();
-extern "C" cudaError_t ComputeDifferenceWithCuda(const unsigned char (*color)[3], const unsigned char *depth, double (*diff)[8], size_t rows, size_t cols);
+extern "C" cudaError_t ComputeDifferenceWithCuda(const unsigned char (*color)[3], const unsigned char *depth, float (*diff)[8], size_t rows, size_t cols);
 
 #endif
